requirements.txt
# Web Framework & API
fastapi>=0.115.0
uvicorn>=0.30.6
streamlit>=1.38.0
pydantic>=2.9.2
python-multipart>=0.0.9
httpx>=0.27.0

# Environment and utilities
python-dotenv>=1.0.1
python-dateutil>=2.9.0

# Web Scraping (used by src/scraper.py and src/scraper_v2.py)
requests>=2.32.3
beautifulsoup4>=4.12.3
lxml>=4.9.2
trafilatura>=1.6.3
playwright>=1.40.0
selectolax>=0.3.21
extruct>=0.16.0
feedparser>=6.0.0

# OpenAI, Pinecone, Instructor
openai>=1.52.0,<2.0.0
instructor>=1.9.0
pinecone>=7.3.0

# Google Cloud Storage
google-cloud-storage>=2.18.2
google-cloud-logging>=3.11.0
google-api-core>=2.19.2
google-auth>=2.26.1,<3.0.0

# Testing
pytest==8.3.4
pytest-asyncio==1.3.0